crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "celery[redis]>=5.5.3",
    "lxml>=6.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.18.1",
]
